+++ src/main/java/com/nhnacademy/inkbridge/backend/domain/MemberAccountRepository.java
@@ -3,8 +3,8 @@ package com.nhnacademy.inkbridge.backend.domain;
 public interface MemberAccountRepository {
 	void save(Member member);
 
-	void update(Integer loginId, EssentialAccountInfo essentialAccountInfo, ProfileInfo profileInfo);
+	void update(Long loginId, EssentialAccountInfo essentialAccountInfo, ProfileInfo profileInfo);
 
-	void delete(Integer loginId);
+	void delete(Long loginId);
 
 }
