+++ src/main/java/com/nhnacademy/inkbridge/backend/storage/MemberAccountJpaRepository.java
@@ -2,5 +2,5 @@ package com.nhnacademy.inkbridge.backend.storage;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberAccountJpaRepository extends JpaRepository<MemberEntity, Integer> {
+public interface MemberAccountJpaRepository extends JpaRepository<MemberEntity, Long> {
 }
